Export import last updated

+++ app/src/main/java/com/wtb/dashTracker/ui/dialog_edit_data_model/dialog_expense/ExpenseViewModel.kt
@@ -25,7 +25,7 @@ import kotlinx.coroutines.flow.Flow
 
 @ExperimentalCoroutinesApi
 class ExpenseViewModel : ListItemViewModel<Expense>() {
-    override fun getItemFlowById(id: Int): Flow<Expense?> =
+    override fun getItemFlowById(id: Long): Flow<Expense?> =
         repository.getExpenseFlowById(id)
 
     val expensePurposes: Flow<List<ExpensePurpose>> = repository.allExpensePurposes
